requirements.txt
sqlalchemy==2.0.31
flask-sqlalchemy-3.1.1
python-dotenv==1.0.1
Flask==3.0.3
Flask-Login==0.6.3
psycopg2-binary==2.9.9
flask-marshmallow==1.2.1
marshmallow-sqlalchemy==1.0.0
